+++ R/estimate_contrasts.R
@@ -39,6 +39,10 @@
 #'     `sequential`, `meandev`, etc., see string-options). Optionally, comparisons
 #'     can be carried out within subsets by indicating the grouping variable
 #'     after a vertical bar ( `|`).
+#'   * A custom function, e.g. `comparison = myfun`, or
+#'     `comparison ~ I(my_fun(x)) | groups`.
+#'   * If contrasts should be calculated (or grouped by) factors, `custom` can
+#'     also be a matrix that specifies factor contrasts (see 'Examples').
 #' @param effectsize Desired measure of standardized effect size, one of
 #' `"emmeans"`, `"marginal"`, or `"boot"`. Default is `NULL`, i.e. no effect
 #' size will be computed.
@@ -116,6 +120,13 @@
 #' # Standardized differences
 #' estimated <- estimate_contrasts(lm(Sepal.Width ~ Species, data = iris))
 #' standardize(estimated)
+#'
+#' # custom factor contrasts - contrasts the average effects of two levels
+#' # against the remaining third level
+#' data(contrast_example, package = "modelbased")
+#' cond_tx <- cbind("no treatment" = c(1, 0, 0), "treatment" = c(0, 0.5, 0.5))
+#' model <- lm(outcome ~ score * tx, data = contrast_example)
+#' estimate_slopes(model, "score", by = "tx", comparison = cond_tx)
 #'
 #' # Other models (mixed, Bayesian, ...)
 #' data <- iris